aggiornamento

+++ scripts/update_ts_files.sh
@@ -1,10 +1,10 @@
-#!/bin/bash
+#!/usr/bin/env bash
 ###########################################################################
 #    update_ts_files.sh
 #    ---------------------
 #    Date                 : July 2007
 #    Copyright            : (C) 2007 by Tim Sutton
-#    Email                : tim dot linfiniti at com
+#    Email                : tim at linfiniti dot com
 ###########################################################################
 #                                                                         #
 #   This program is free software; you can redistribute it and/or modify  #
@@ -21,36 +21,27 @@
 # Note the .pro file must NOT be named qgis.pro as this
 # name is reserved for the Windows qmake project file
 
+echo "deprecated - use push_ts.sh and pull_ts.sh" >&2
+
 set -e
 
 cleanup() {
-	if [ -f i18n/python_ts.tar ]; then
-		tar -xf i18n/python_ts.tar
-	fi
 	if [ -f i18n/qgis_ts.tar ]; then
 		echo Restoring excluded translations
 		tar -xf i18n/qgis_ts.tar
 	fi
 
 	echo Removing temporary files
-	perl -i.bak -ne 'print unless /^\s+<location.*python-i18n\.cpp.*$/;' i18n/qgis_*.ts
+	perl -i.bak -ne 'print unless /^\s+<location.*(python-i18n|_texts)\.cpp.*$/;' i18n/qgis_*.ts
 	for i in \
 		python/python-i18n.{ts,cpp} \
 		python/plugins/*/python-i18n.{ts,cpp} \
 		i18n/qgis_*.ts.bak \
 		src/plugins/grass/grasslabels-i18n.cpp \
 		i18n/qgis_ts.tar \
-		i18n/python_ts.tar \
 		qgis_ts.pro
 	do
 		[ -f "$i" ] && rm "$i"
-	done
-
-	for i in \
-		src/plugins/plugin_template/plugingui.cpp \
-		src/plugins/plugin_template/plugin.cpp
-	do
-		[ -f "$i.save" ] && mv "$i.save" "$i"
 	done
 
 	trap "" EXIT
@@ -58,19 +49,19 @@ cleanup() {
 
 PATH=$QTDIR/bin:$PATH
 
-if type qmake-qt4 >/dev/null 2>&1; then
-	QMAKE=qmake-qt4
+if type qmake-qt5 >/dev/null 2>&1; then
+	QMAKE=qmake-qt5
 else
 	QMAKE=qmake
 fi
 
-if ! type pylupdate4 >/dev/null 2>&1; then
-      echo "pylupdate4 not found"
+if ! type pylupdate5 >/dev/null 2>&1; then
+      echo "pylupdate5 not found"
       exit 1
 fi
 
-if type lupdate-qt4 >/dev/null 2>&1; then
-	LUPDATE=lupdate-qt4
+if type lupdate-qt5 >/dev/null 2>&1; then
+	LUPDATE=lupdate-qt5
 else
 	LUPDATE=lupdate
 fi
@@ -102,36 +93,40 @@ done
 
 trap cleanup EXIT
 
-tar --remove-file -cf i18n/python_ts.tar $(find python -name "*.ts")
 if [ "$exclude" != "--exclude i18n/qgis_en.ts" -o -n "$add" ]; then
   echo Saving excluded translations
   tar $fast -cf i18n/qgis_ts.tar i18n/qgis_*.ts $exclude
 fi
 
+builddir=$1
+if [ -d "$builddir" ]; then
+	echo Build directory not found
+	exit 1
+fi
+
+if [ ! -f "$builddir/src/core/qgsexpression_texts.cpp" ]; then
+	echo Generated help files not found
+	exit 1
+fi
+
 echo Updating python translations
-cd python
-pylupdate4 utils.py {console,pyplugin_installer}/*.{py,ui} -ts python-i18n.ts
-perl ../scripts/ts2cpp.pl python-i18n.ts python-i18n.cpp
-rm python-i18n.ts
-cd ..
+(
+	cd python
+	pylupdate5 utils.py {console,pyplugin_installer}/*.{py,ui} -ts python-i18n.ts
+	perl ../scripts/ts2cpp.pl python-i18n.ts python-i18n.cpp
+	rm python-i18n.ts
+)
 for i in python/plugins/*/CMakeLists.txt; do
 	cd ${i%/*}
-	pylupdate4 $(find . -name "*.py" -o -name "*.ui") -ts python-i18n.ts
+	pylupdate5 $(find . -name "*.py" -o -name "*.ui") -ts python-i18n.ts
 	perl ../../../scripts/ts2cpp.pl python-i18n.ts python-i18n.cpp
 	rm python-i18n.ts
 	cd ../../..
 done
 echo Updating GRASS module translations
 perl scripts/qgm2cpp.pl >src/plugins/grass/grasslabels-i18n.cpp
-mv src/plugins/plugin_template/plugingui.cpp src/plugins/plugin_template/plugingui.cpp.save
 echo Creating qmake project file
-for i in \
-	src/plugins/plugin_template/plugingui.cpp \
-	src/plugins/plugin_template/plugin.cpp
-do
-	[ -f "$i" ] && mv "$i" "$i.save"
-done
-$QMAKE -project -o qgis_ts.pro -nopwd src python i18n
+$QMAKE -project -o qgis_ts.pro -nopwd src python i18n "$builddir/src/core/qgsexpression_texts.cpp"
 if [ -n "$add" ]; then
 	for i in $add; do
 		echo "Adding translation for $i"
